+++ nl_banadir/banadir_customization_reports/report/accounts_payable_summary_multicurrency/accounts_payable_summary_multicurrency.json
@@ -1,0 +1,35 @@
+{
+ "add_total_row": 1,
+ "columns": [],
+ "creation": "2024-07-04 13:46:25.354424",
+ "disabled": 0,
+ "docstatus": 0,
+ "doctype": "Report",
+ "filters": [],
+ "idx": 0,
+ "is_standard": "Yes",
+ "letterhead": null,
+ "modified": "2024-07-04 13:50:33.934133",
+ "modified_by": "Administrator",
+ "module": "Banadir Customization-Reports",
+ "name": "Accounts Payable Summary MultiCurrency",
+ "owner": "Administrator",
+ "prepared_report": 0,
+ "ref_doctype": "Purchase Invoice",
+ "report_name": "Accounts Payable Summary MultiCurrency",
+ "report_type": "Script Report",
+ "roles": [
+  {
+   "role": "Accounts User"
+  },
+  {
+   "role": "Purchase User"
+  },
+  {
+   "role": "Accounts Manager"
+  },
+  {
+   "role": "Auditor"
+  }
+ ]
+}
