+++ nl_banadir/banadir_customization_reports/report/accounts_payable_summary_multicurrency/accounts_payable_summary_multicurrency.py
@@ -1,0 +1,17 @@
+# Copyright (c) 2024, Navari Ltd and contributors
+# For license information, please see license.txt
+
+# import frappe
+
+
+from nl_banadir.banadir_customization_reports.report.accounts_receivable_summary_multicurrency.accounts_receivable_summary_multicurrency import (
+	AccountsReceivableSummary,
+)
+
+
+def execute(filters=None):
+	args = {
+		"account_type": "Payable",
+		"naming_by": ["Buying Settings", "supp_master_name"],
+	}
+	return AccountsReceivableSummary(filters).run(args)
