--- nl_banadir/banadir_customization_reports/report/accounts_receivable/accounts_receivable.json
@@ -1,26 +1,0 @@
-{
- "add_total_row": 1, 
- "apply_user_permissions": 1, 
- "creation": "2013-04-16 11:31:13", 
- "disabled": 0, 
- "docstatus": 0, 
- "doctype": "Report", 
- "idx": 3, 
- "is_standard": "Yes", 
- "modified": "2017-03-06 05:52:06.235584", 
- "modified_by": "Administrator", 
- "module": "Accounts", 
- "name": "Accounts Receivable", 
- "owner": "Administrator", 
- "ref_doctype": "Sales Invoice", 
- "report_name": "Accounts Receivable", 
- "report_type": "Script Report", 
- "roles": [
-  {
-   "role": "Accounts Manager"
-  }, 
-  {
-   "role": "Accounts User"
-  }
- ]
-}
